MCPJVM-395: move Probe and Artifact Management actions into Feature Modules
