+++ test/tools/spring/artifact-management-domain.test.ts
@@ -3,7 +3,7 @@ const fs = require("node:fs");
 const path = require("node:path");
 const test = require("node:test");
 
-const { artifactManagementDomain } = require("@/tools/core/artifact_management/domain");
+const { artifactManagementDomain } = require("@tools-feature-artifact-management/domain");
 
 function createTestTempDir(prefix: string): string {
   const base = path.join(process.cwd(), "test", ".tmp");
