+++ tools/features/artifact-management/actions/run_result.action.ts
@@ -1,10 +1,10 @@
 import { promises as fs } from "node:fs";
 import path from "node:path";
 import { resolveRegressionRunDirAbs } from "@tools-regression-execution-plan-spec/regression_results_report.util";
-import type { ArtifactActionContext, ArtifactActionRequest, ArtifactActionResult } from "@/tools/core/artifact_management/actions/types";
-import { buildFailClosedArtifactResponse, okArtifactResponse } from "@/tools/core/artifact_management/shared/fail_closed.util";
-import { readJsonFile } from "@/tools/core/artifact_management/shared/json_io.util";
-import { resolveProjectName } from "@/tools/core/artifact_management/shared/project_resolution.util";
+import type { ArtifactActionContext, ArtifactActionRequest, ArtifactActionResult } from "@tools-feature-artifact-management/actions/types";
+import { buildFailClosedArtifactResponse, okArtifactResponse } from "@tools-feature-artifact-management/shared/fail_closed.util";
+import { readJsonFile } from "@tools-feature-artifact-management/shared/json_io.util";
+import { resolveProjectName } from "@tools-feature-artifact-management/shared/project_resolution.util";
 
 function asStringArray(value: unknown): string[] {
   if (!Array.isArray(value)) return [];
