+++ tools/features/artifact-management/shared/project_resolution.util.ts
