+++ tools/features/probe/index.ts
@@ -1,2 +1,3 @@
-/** Public Probe Feature Module boundary. Active implementation migration is a later story. */
+export { createProbeDomain, executeProbeAction } from "./domain";
+export type { ProbeActionRequest, ProbeDomainConfig } from "./domain";
 export type ProbeFeatureModule = "probe";
