+++ tools/transport/tools-mcp-server/src/tools/core/artifact_management/handler.ts
@@ -2,8 +2,7 @@ import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import type { ProbeRegistrySummary } from "@/config/probe-registry";
 import { ArtifactManagementRequestSchema } from "@/models/inputs/artifact_management";
 import { ARTIFACT_MANAGEMENT_TOOL } from "@/tools/core/artifact_management/contract";
-import { artifactManagementDomain } from "@/tools/core/artifact_management/domain";
-import { buildFailClosedArtifactResponse } from "@/tools/core/artifact_management/shared/fail_closed.util";
+import { artifactManagementDomain, buildFailClosedArtifactResponse } from "@tools-feature-artifact-management";
 
 export type ArtifactManagementHandlerDeps = {
   workspaceRootAbs: string;
